Alvaro's return o, the previous change
OK then, Alvaro. I’ll do that change.

The bottom line is that this draft addresses a brown field and will not change the arrangements as deployed, so it does not really matter what we say. Maybe it does not hurt to mention that as well does it?

The result would be:
“
   It is worth noting that with RPL <xref target='RFC6550'/>, every node in the
   LLN that is RPL-aware can inject any RPL-based attack in the network.
   This document applies typically to an existing deployment and does not change
   its security requirements and operations.
   It is assumed that the security mechanisms as defined for RPL are followed.
“

This is undoing what Stewart recommended (uppercase). But Stewart is CC so he can come back to us.
In the meantime I’ll do as you suggest in github. Do you prefer I publish already?

Take care,

Pascal

From: Alvaro Retana <aretana.ietf@gmail.com>
Sent: jeudi 27 août 2020 22:46
To: Pascal Thubert (pthubert) <pthubert@cisco.com>
Cc: rtg-dir@ietf.org; Stewart Bryant <stewart.bryant@gmail.com>
Subject: Re: [RTG-DIR] Rtgdir last call review of draft-ietf-roll-turnon-rfc8138-10

Hi!

If we want to close the gap, it should probably be in a document that Updates rfc6550.   Or maybe the WG wants to wait until RPLv2 and do it there…

:-)

Take care!

Alvaro.

On August 27, 2020 at 4:38:24 PM, Pascal Thubert (pthubert) (pthubert@cisco.com) wrote:
Hello Alvaro

The point is if the security measures in RPL are not turned on, it’s either L2 security or none at all.

Shouldn’t we mandate L2 security in that case?

The quote from NP DAO is imprecise in that the security section in RPL only discusses the secured modes.  RFC 6550 says nothing in case of unsecured because we expected secured. It fails to recommend L2 security. The NPDAO text this inherits that gap.

The real world decided otherwise  and uses L2 security ubiquitously. Shouldn’t we reflect that ? Saying that we expect RPL security is a denial of the world around us.

So I’m good with your proposal but I feel the gap should be closed somewhere.  Agreed this is but a side spec, but then where should that be?

Take care,

Pascal